Split mn_22EC out of mnmain and link both
The DOL's mnmain slice holds two literal pools: 0x804DBD88..0x804DBDDC is
used only by the functions up to mn_8022DDA8_OnEnter, and 0x804DBDE0 onward
only by mn_8022EC18 and later, with 0.0f appearing in both. One translation
unit cannot emit a constant twice, so the object was assembled from two
sources; splitting at mn_8022EC18 reproduces both pools.

main.dol is byte-identical and dtk dol diff is clean.
